+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - CSS issue in documentation rendering for <pre></pre> block in dark theme
 - Template issue on documentation when applying jinja filters on None
 - Remove superusers from "list_approvers". They still can approve but are not displayed in the list.
+- When editing an ApprovalWorkflow, it was possible to use a scope already assigned to another workflow.
 
 ## Enhancement